move google-auth.js to common/api/authentication
