+++ extension/js/common/api/authentication/google-auth.ts
@@ -2,23 +2,23 @@
 
 'use strict';
 
-import { Str, Url } from '../../../core/common.js';
-import { FLAVOR, GOOGLE_OAUTH_SCREEN_HOST, OAUTH_GOOGLE_API_HOST } from '../../../core/const.js';
-import { ApiErr } from '../../shared/api-error.js';
-import { Api, ApiCallContext } from './../../shared/api.js';
+import { Str, Url } from '../../core/common.js';
+import { FLAVOR, GOOGLE_OAUTH_SCREEN_HOST, OAUTH_GOOGLE_API_HOST } from '../../core/const.js';
+import { ApiErr } from '../shared/api-error.js';
+import { Api, ApiCallContext } from '../shared/api.js';
 
-import { Bm, GoogleAuthWindowResult$result } from '../../../browser/browser-msg.js';
-import { Buf } from '../../../core/buf.js';
-import { InMemoryStoreKeys } from '../../../core/const.js';
-import { OAuth2 } from '../../../oauth2/oauth2.js';
-import { Catch } from '../../../platform/catch.js';
-import { AcctStore, AcctStoreDict } from '../../../platform/store/acct-store.js';
-import { InMemoryStore } from '../../../platform/store/in-memory-store.js';
-import { AccountServer } from '../../account-server.js';
-import { ExternalService } from '../../account-servers/external-service.js';
-import { GoogleAuthErr } from '../../shared/api-error.js';
-import { GmailRes } from './gmail-parser.js';
-import { Assert, AssertError } from '../../../assert.js';
+import { Bm, GoogleAuthWindowResult$result } from '../../browser/browser-msg.js';
+import { Buf } from '../../core/buf.js';
+import { InMemoryStoreKeys } from '../../core/const.js';
+import { OAuth2 } from '../../oauth2/oauth2.js';
+import { Catch } from '../../platform/catch.js';
+import { AcctStore, AcctStoreDict } from '../../platform/store/acct-store.js';
+import { InMemoryStore } from '../../platform/store/in-memory-store.js';
+import { AccountServer } from '../account-server.js';
+import { ExternalService } from '../account-servers/external-service.js';
+import { GoogleAuthErr } from '../shared/api-error.js';
+import { GmailRes } from '../email-provider/gmail/gmail-parser.js';
+import { Assert, AssertError } from '../../assert.js';
 
 /* eslint-disable @typescript-eslint/naming-convention */
 type GoogleAuthTokensResponse = {
